refactor: extract Mysql and its child class to module
